third_party/blink/web_tests/fast/css/style-element-no-specifier-warning-expected.txt

CONSOLE WARNING: <style type=module> has no `specifier` value
This is a testharness.js-based test.
Harness: the test ran to completion.
